requirements.txt
amqp==5.1.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==6.0.0
bs4==0.0.1
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==4.0.0
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
colorama==0.4.4
comm==0.1.2
constantly==15.1.0
cryptography==38.0.1
cssselect==1.1.0
cycler==0.10.0
daphne==3.0.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.2
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==4.1.3
django-advanced-filters==2.0.0
django-allauth==0.46.0
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-ckeditor==6.2.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-daterange-filter==1.3.0
django-daterangefilter==1.0.0
django-datetimepicker==3.14
django-dotenv==1.4.2
django-environ==0.8.1
django-filter==2.4.0
django-fixture-magic==0.1.5
django-heroku==0.3.1
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-mysql==4.2.0
django-rest-framework==0.1.0
django-restframework==0.0.1
django-sendsms==0.4
django-timezone-field==5.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-instamojo==0.0.1
drfaddons==0.1.4
environ==1.0
executing==1.2.0
fake-useragent==0.1.11
Faker==10.0.0
fastjsonschema==2.16.3
feedparser==6.0.10
filelock==3.0.12
Flask==2.0.2
fqdn==1.5.1
future==0.18.2
gunicorn==20.1.0
heroku==0.1.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==5.0.0
incremental==21.3.0
instamojo-wrapper==1.1
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==3.0.3
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.3.1
kombu==5.2.4
lxml==4.9.1
mailchimp-marketing==3.0.74
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mibian==0.1.3
mistune==2.0.5
msgpack==1.0.4
multitasking==0.0.11
mysql-connector-python==8.2.0
mysqlclient==2.1.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
node==1.2.1
nodejs==0.1.1
notebook==6.5.3
notebook_shim==0.2.2
npm==0.1.1
numpy==1.21.1
oauthlib==3.1.1
odict==1.9.0
optional-django==0.1.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
plumber==1.7
prometheus-client==0.16.0
prompt-toolkit==3.0.31
protobuf==4.21.12
psutil==5.9.4
psycopg2==2.9.1
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==8.2.2
Pygments==2.14.0
PyJWT==2.1.0
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==2.4.7
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.19.3
python-crontab==2.6.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2021.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
rdflib==6.2.0
rdflib-jsonld==0.6.2
redis==4.3.4
requests==2.26.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scipy==1.7.0
Send2Trash==1.8.0
service-identity==21.1.0
sgmllib3k==1.0.0
simplejson==3.17.6
six==1.16.0
smartapi-python==1.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
stack-data==0.6.2
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
toposort==1.7
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
uri-template==1.2.0
urllib3==1.26.7
vine==5.0.0
virtualenv==20.7.0
virtualenvwrapper-win==1.2.6
w3lib==2.0.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.3
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.14.1
yahoo-fin==0.8.9.1
yfinance==0.1.75
zipp==3.8.1
zope.component==6.0
zope.deferredimport==4.4
zope.deprecation==5.0
zope.event==4.6
zope.hookable==5.4
zope.interface==5.5.0
zope.lifecycleevent==4.4
zope.proxy==5.0.0